Change rating column from int to float

+++ app/models/rating.py
@@ -10,7 +10,7 @@ class Rating(Base):
 
     user_id: Mapped[int] = mapped_column(ForeignKey("user.id"), primary_key=True)
     book_id: Mapped[int] = mapped_column(ForeignKey("book.id"), primary_key=True)
-    rating: Mapped[int]
+    rating: Mapped[float]
     rating_text: Mapped[str | None]
     date_created: Mapped[datetime] = mapped_column(server_default=func.now())
     date_updated: Mapped[datetime] = mapped_column(server_default=func.now(), onupdate=func.now())
